Added the ability to use `ValueReference`.
Now it is possible to assign serializable classes to localizations.
This functionality should simplify the use of localizations, as well as slightly improve performance when getting objects of type `Component`, because it will not be necessary to deserialize it from the configuration every time.
The new functionality applies only to file formats supported by Sponge.