Version 5.0
The `SerializedItemStackJsonNbt` and `SerializedItemStackPlainNBT` classes are now marked as deprecated and will be removed in the future. They have been replaced by a single universal class `SerializedItemStack`.
Tracking of changes in localizations has been improved. Fixed a bug that made it necessary to save the localization again to reload it successfully.
The FileUtils class has been added to the API. Currently it only allows to get file extension. The functionality can be expanded in the future.
A small optimization of the `LocaleReference` interface has been done.
Other minor fixes have been made.